rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-anthropic>=0.3.0",
    "langchain-huggingface>=0.1.0",
    "langchain-qdrant>=0.2.0",
    "qdrant-client>=1.12.0",
    "pymupdf4llm>=0.0.17",
    "fastembed>=0.4.0",
    "tiktoken>=0.8.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]
